requirements.txt
accelerate==0.32.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anthropic==0.26.0
anyio==3.7.1
appnope @ file:///Users/cbousseau/work/recipes/ci_py311/appnope_1677917710869/work
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.3
attrs==23.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beartype==0.14.1
better-abc==0.0.3
black==23.10.1
blinker==1.8.2
cachetools==5.3.3
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
comm @ file:///Users/cbousseau/work/recipes/ci_py311/comm_1677919149446/work
contourpy==1.2.0
cycler==0.12.1
datasets==2.14.6
debugpy @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_563_nwtkoc/croot/debugpy_1690905063850/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dill==0.3.7
distro==1.8.0
docker-pycreds==0.4.0
einop==0.0.1
einops==0.7.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fancy-einsum==0.0.3
filelock==3.13.1
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
huggingface-hub==0.23.4
idna==3.4
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1698244280508/work
ipython @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_6599f73fa7/croot/ipython_1694181355402/work
jaxtyping==0.2.31
jedi @ file:///Users/cbousseau/work/recipes/ci_py311_2/jedi_1678994967789/work
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_58w2siozyz/croot/jupyter_client_1699455907045/work
jupyter_core @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_782yoyc_98/croot/jupyter_core_1698937318631/work
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline @ file:///Users/cbousseau/work/recipes/ci_py311/matplotlib-inline_1677918241899/work
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio @ file:///Users/cbousseau/work/recipes/ci_py311/nest-asyncio_1677912430289/work
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
openai==1.1.1
packaging @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_6dm6d4jd_t/croot/packaging_1693575176524/work
pandas==2.1.2
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathspec==0.11.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==10.1.0
platformdirs @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a8u4fy8k9o/croot/platformdirs_1692205661656/work
prettytable==3.9.0
prompt-toolkit @ file:///Users/cbousseau/work/recipes/ci_py311/prompt-toolkit_1677918689663/work
protobuf==5.27.2
psutil @ file:///Users/cbousseau/work/recipes/ci_py311_2/psutil_1678995687212/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==14.0.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.9.1
Pygments @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_29bs9f_dh9/croot/pygments_1684279974747/work
pyparsing==3.1.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_23n9bfwjq5/croot/pyzmq_1686601381911/work
referencing==0.35.1
regex==2023.10.3
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.11.3
sentencepiece==0.2.0
sentry-sdk==2.7.1
setproctitle==1.3.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sniffio==1.3.0
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
streamlit==1.36.0
-e git+https://github.com/JayThibs/supervising-ais-improving-ais.git@1326a555f5afed302730a14677fc87ca792520e3#egg=supervising_ais
sympy==1.12
tenacity==8.2.3
terminaltables==3.1.10
threadpoolctl==3.2.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
tornado @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_3a5nrn2jeh/croot/tornado_1696936974091/work
tqdm==4.66.1
traitlets @ file:///Users/cbousseau/work/recipes/ci_py311/traitlets_1677911650502/work
transformer-lens==2.1.0
transformers==4.42.3
treelib==1.7.0
typeguard==2.13.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
wandb==0.17.4
wcwidth==0.2.9
wordcloud==1.9.3
xxhash==3.4.1
yarl==1.9.2
